Audio Policy: fix error when speaker strategy is forced during a voice call
In the commit ab7dd802ba82363f11c96931301fbe0fc1f8432b, when speaker strategy is forced
during a voice call, the audio devices used will be all the current available devices.
For example if voice call is in Earpiece mode and phone wants to switch to speaker,
at the end earpiece and speaker will be enabled.

Change-Id: Id4c28bff35e803314d30e09f3f4d35f291265197
Signed-off-by: Francois Mazard <f-mazard@ti.com>