OMAP4: Call standby when disabling FM Rx
After FM playback, audio was not able to be played properly because
the device was being to 0 when calling close.  Instead, by calling
standby, this device is never set to 0.

Resolves DR:  OMAPS00232464

Signed-off-by: Ricky Bhatia <a0866433@ti.com>
Signed-off-by: Chris Kelly <c-kelly@ti.com>

Patch set 2: A simple cherry-pick of the patch onto the current
tree.  No changes made.  Taking approvals, verifications and build
test from patch set 1.
Signed-off-by: Jean Johnson <jean-johnson@ti.com>
Change-Id: Ifc61a3805e15c967358769894452cbea1c27762f