Remove partial wake locks
The power manager service allows applications to register
with it and set wake_locks as required. This allows applications
to keep the screen on if required or continue while the screen is off.
Since the power policy is maintained in applications,
there is no need to have a redundant wake lock held in the audio hal.

This is a fix for OMAPS00232171.
The problem was root-caused to the PowerManagerService wake lock
being released before the AudioOutLock. This condition should
not happen and caused unexpected behavior with some apks.

Change-Id: I1dcbfd5afdb8f56c654bf67e9ddff5ab10357207
Signed-off-by: Chris Kelly <c-kelly@ti.com>