Revert "Remove partial wake locks"
This reverts commit 4a6e7dca3d3f7a8a195e3b9c7805f9e0ba9daf4e.
This was initially intended to fix strange behavior from certain
game apps where audio continued to play after suspend/pwr button
was pushed. It is typically expected to end the game audio if the
screen is off.

However, it turns out that the games with this behavior are
not correctly implemented somehow and we require the wake locks
here.